requirements.txt
openai-whisper
PyQt5
